discord: Use Index data for price searches
Determine if a card exists by looking at the Index, so that it's
possible to heavily simplify searches for Sellers and Vendors.

Display the Index data to the list of results as first element,
so that it appears at the top and the other prices follow below.

Support 'i' for index redirects.