init: Fix loading scrapers with the same name
The scraper slices are sorted by name, so that their positions can always
be predictable, however when there are two identical ones (but with
different shorthand) their resulting order depends on the map position,
which is random and unpredictable.

This lead to some scraper overwriting the one with the same name,
since one would randomly appear earlier than the other. Now the scapers
with the same name are sorted by their shorthand, which is unique, and
avoids this kind of collisions.